API 작업 전 결합
