+++ Brandingdong/Brandingdong/API/Service.swift
@@ -8,6 +8,11 @@
 
 import UIKit
 
+struct Image : Codable {
+  var images : String
+}
+
+
 struct Service {
   
   static func signUpUser(username: String, email: String, password1: String, password2: String, phonenumber: String) {
@@ -69,32 +74,4 @@ struct Service {
     task.resume()
   }
   
-  //  static func signInUser(username: String, password: String) {
-  //    print ("서비스")
-  //    let loginUrl = "http://52.78.75.94/auth/login/"
-  //    guard let url = URL(string: loginUrl) else { return }
-  //    print ("서비스1")
-  //    let task = URLSession.shared.dataTask(with: url) { (data, response, error) in
-  //      guard error == nil else { return print ("error : ", error!.localizedDescription)}
-  //      print ("서비스2")
-  //      let responsea = response as? HTTPURLResponse
-  //      print (responsea!.statusCode)
-  //      guard let response = response as? HTTPURLResponse,
-  //        (200..<406).contains(response.statusCode) else { return }
-  //      print ("서비스3")
-  //      guard let data = data else { return }
-  //      print ("서비스4")
-  //      var userResult: SignInData?
-  //
-  //      do {
-  //        userResult = try JSONDecoder().decode(SignInData.self, from: data)
-  //        print("userResult?.username : ", userResult?.username)
-  //        print("userResult?.password : ", userResult?.password)
-  //        print ("서비스4")
-  //      } catch {
-  //        print ("failed to convert error : ", error.localizedDescription)
-  //      }
-  //    }
-  //    task.resume()
-  //  }
 }
