+++ Brandingdong/Brandingdong/SupportFile/Assets.xcassets/주문정보.imageset/Contents.json
@@ -1,6 +1,7 @@
 {
   "images" : [
     {
+      "filename" : "주문정보.jpeg",
       "idiom" : "universal",
       "scale" : "1x"
     },
@@ -9,7 +10,6 @@
       "scale" : "2x"
     },
     {
-      "filename" : "주문정보.jpeg",
       "idiom" : "universal",
       "scale" : "3x"
     }
